// assets/src/promotion_carousel_page/interface.ts
export interface PromotionCarouselPageProps {
    link: string;
    title: string;
    content: string;
    mediaId: number;
    mediaUrl: string;
    index?: number;
    contentUrl: string;
}

export interface MediaViewProps {
    medialURL: string;
    mediaId: number;
}

export interface PromotionCarouselPageWrapperProps {
    link: string;
    title: string;
    content: string;
    mediaId: number;
    mediaUrl: string;
    contentUrl: string;
}